Move decision to build a tunnel or not to the shell script

+++ tunnel.sh
@@ -6,6 +6,7 @@ ret=0
 
 if [ -z "$MPID" ] ; then
   if [ -n "$TUNNEL_DEBUG" ] ; then
+    exec 2>/tmp/t1
     set -x
     env >&2
   fi
@@ -23,33 +24,53 @@ if [ -z "$MPID" ] ; then
   export TARGET_PORT="`echo $query | sed -e 's/^.*\"target_port\": *\"//' -e 's/\".*$//g'`"
   export GATEWAY_HOST="`echo $query | sed -e 's/^.*\"gateway_host\": *\"//' -e 's/\".*$//g'`"
   export GATEWAY_PORT="`echo $query | sed -e 's/^.*\"gateway_port\": *\"//' -e 's/\".*$//g'`"
+  export GATEWAY_USER="`echo $query | sed -e 's/^.*\"gateway_user\": *\"//' -e 's/\".*$//g'`"
   export SHELL_CMD="`echo $query | sed -e 's/^.*\"shell_cmd\": *\"//' -e 's/\",.*$//g' -e 's/\\\"/\"/g'`"
   export SSH_TUNNEL_CHECK_SLEEP="`echo $query | sed -e 's/^.*\"ssh_tunnel_check_sleep\": *\"//' -e 's/\",.*$//g' -e 's/\\\"/\"/g'`"
+  export CREATE="`echo $query | sed -e 's/^.*\"create\": *\"//' -e 's/\",.*$//g' -e 's/\\\"/\"/g'`"
 
-  echo "{ \"host\": \"$LOCAL_HOST\" }"
-  p=`ps -p $PPID -o "ppid="`
-  clog=`mktemp`
-  nohup timeout $TIMEOUT $SHELL_CMD "$ABSPATH/tunnel.sh" $p <&- >&- 2>$clog &
-  CPID=$!
-  # A little time for the SSH tunnel process to start or fail
-  sleep 3
-  # If the child process does not exist anymore after this delay, report failure
-  if ! ps -p $CPID >/dev/null 2>&1 ; then
-    echo "Child process ($CPID) failure - Aborting" >&2
-    echo "Child diagnostics follow:" >&2
-    cat $clog >&2
-    rm -f $clog
-    ret=1
+  if [ "X$CREATE" = X -o "X$GATEWAY_HOST" = X ] ; then
+    # No tunnel - connect directly to target host
+    do_tunnel=''
+    cnx_host="$TARGET_HOST"
+    cnx_port="$TARGET_PORT"
+  else
+    do_tunnel='y'
+    cnx_host="$LOCAL_HOST"
+    cnx_port="$LOCAL_PORT"
   fi
-  rm -f $clog
+
+  echo "{ \"host\": \"$cnx_host\",  \"port\": \"$cnx_port\" }"
+
+  if [ -n "$do_tunnel" ] ; then
+    p=`ps -p $PPID -o "ppid="`
+    clog=`mktemp`
+    nohup timeout $TIMEOUT $SHELL_CMD "$ABSPATH/tunnel.sh" $p <&- >&- 2>$clog &
+    CPID=$!
+    # A little time for the SSH tunnel process to start or fail
+    sleep 3
+    # If the child process does not exist anymore after this delay, report failure
+    if ! ps -p $CPID >/dev/null 2>&1 ; then
+      echo "Child process ($CPID) failure - Aborting" >&2
+      echo "Child diagnostics follow:" >&2
+      cat $clog >&2
+      rm -f $clog
+      ret=1
+    fi
+    rm -f $clog
+  fi
 else
   #------ Child
   if [ -n "$TUNNEL_DEBUG" ] ; then
+    exec 2>/tmp/t2
     set -x
     env >&2
   fi
 
-  $SSH_CMD -N -L localhost:$LOCAL_PORT:$TARGET_HOST:$TARGET_PORT -p $GATEWAY_PORT $GATEWAY_HOST &
+  gw="$GATEWAY_HOST"
+  [ "X$GATEWAY_USER" = X ] || gw="$GATEWAY_USER@$GATEWAY_HOST"
+
+  $SSH_CMD -N -L $LOCAL_HOST:$LOCAL_PORT:$TARGET_HOST:$TARGET_PORT -p $GATEWAY_PORT $gw &
   CPID=$!
   
   sleep $SSH_TUNNEL_CHECK_SLEEP
